pyproject.toml
[project]
name = "ai-meeting-secretary"
version = "0.1.0"
description = "Foundation v1 for AI Meeting Secretary"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.0",
    "celery>=5.5.0",
    "faster-whisper>=1.1.0",
    "fastapi>=0.115.0",
    "openai>=1.107.0",
    "pyannote.audio>=3.3.2",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.10.0",
    "psycopg[binary]>=3.2.0",
    "python-multipart>=0.0.20",
    "redis>=6.0.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.35.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
]
